Rename airflow.sensors.date_time_sensor to airflow.sensors.date_time
