+++ airflow/sensors/date_time_sensor.py
@@ -15,59 +15,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use `airflow.sensors.date_time`."""
 
-import datetime
-from typing import Dict, Union
+import warnings
 
-from airflow.sensors.base import BaseSensorOperator
-from airflow.utils import timezone
-from airflow.utils.decorators import apply_defaults
+# pylint: disable=unused-import
+from airflow.sensors.date_time import DateTimeSensor  # noqa
 
+warnings.warn(
+    "This module is deprecated. Please use `airflow.sensors.date_time`.", DeprecationWarning, stacklevel=2
+)
 
-class DateTimeSensor(BaseSensorOperator):
-    """
-    Waits until the specified datetime.
-
-    A major advantage of this sensor is idempotence for the ``target_time``.
-    It handles some cases for which ``TimeSensor`` and ``TimeDeltaSensor`` are not suited.
-
-    **Example** 1 :
-        If a task needs to wait for 11am on each ``execution_date``. Using
-        ``TimeSensor`` or ``TimeDeltaSensor``, all backfill tasks started at
-        1am have to wait for 10 hours. This is unnecessary, e.g. a backfill
-        task with ``{{ ds }} = '1970-01-01'`` does not need to wait because
-        ``1970-01-01T11:00:00`` has already passed.
-
-    **Example** 2 :
-        If a DAG is scheduled to run at 23:00 daily, but one of the tasks is
-        required to run at 01:00 next day, using ``TimeSensor`` will return
-        ``True`` immediately because 23:00 > 01:00. Instead, we can do this:
-
-        .. code-block:: python
-
-            DateTimeSensor(
-                task_id='wait_for_0100',
-                target_time='{{ next_execution_date.tomorrow().replace(hour=1) }}',
-            )
-
-    :param target_time: datetime after which the job succeeds. (templated)
-    :type target_time: str or datetime.datetime
-    """
-
-    template_fields = ("target_time",)
-
-    @apply_defaults
-    def __init__(self, *, target_time: Union[str, datetime.datetime], **kwargs) -> None:
-        super().__init__(**kwargs)
-        if isinstance(target_time, datetime.datetime):
-            self.target_time = target_time.isoformat()
-        elif isinstance(target_time, str):
-            self.target_time = target_time
-        else:
-            raise TypeError(
-                "Expected str or datetime.datetime type for target_time. Got {}".format(type(target_time))
-            )
-
-    def poke(self, context: Dict) -> bool:
-        self.log.info("Checking if the time (%s) has come", self.target_time)
-        return timezone.utcnow() > timezone.parse(self.target_time)
